# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ApplicationList": {
        "properties": {
          "has_more": {
            "description": "True if another page exists; fetch it with offset + limit.",
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ApplicationInfo"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "limit",
          "offset",
          "has_more"
        ],
        "title": "ApplicationList",
        "type": "object"
      },
      "ApplicationStatus": {
        "enum": [
          "applied",
          "assessment",
          "interview",
          "offer",
          "rejected",
          "withdrawn"
        ],
        "title": "ApplicationStatus",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobList": {
        "properties": {
          "has_more": {
            "description": "True if another page exists; fetch it with offset + limit.",
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "limit",
          "offset",
          "has_more"
        ],
        "title": "JobList",
        "type": "object"
      },
...
      "Source": {
        "enum": [
          "linkedin",
          "himalayas",
          "jooble"
        ],
        "title": "Source",
        "type": "string"
      },
...
      "WorkType": {
        "enum": [
          "Remote",
          "Remote?",
          "Hybrid",
          "On-site"
        ],
        "title": "WorkType",
        "type": "string"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "description": "The value of JOBPILOT_API_KEY on the server.",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Read-only access to JobPilot's job-market data: internship and entry-level tech jobs collected daily, each scored against the owner's CV, plus which skills are in demand. Every data endpoint needs an X-API-Key header; /health doesn't.",
    "title": "JobPilot API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/applications": {
      "get": {
        "description": "Returns one page of tracked applications, most recently applied first, with their current status. Use open=true for the ones still in progress (not offer, rejected or withdrawn).",
        "operationId": "list_applications",
        "parameters": [
          {
            "description": "Only applications with this status.",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ApplicationStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only applications with this status.",
              "title": "Status"
            }
          },
          {
            "description": "true: only applications still in progress (not offer, rejected or withdrawn).",
            "in": "query",
            "name": "open",
            "required": false,
            "schema": {
              "default": false,
              "description": "true: only applications still in progress (not offer, rejected or withdrawn).",
              "title": "Open",
              "type": "boolean"
            }
          },
          {
            "description": "Page size (1-100).",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Page size (1-100).",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "How many applications to skip, for paging (0-10000).",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "How many applications to skip, for paging (0-10000).",
              "maximum": 10000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List the owner's job applications, newest first"
      }
    },
...
    "/jobs": {
      "get": {
        "description": "Returns one page of job summaries (no description text), sorted by match_score (highest first), then most recently seen. All filters are optional and combine with AND. Use get_job for the full posting.",
        "operationId": "list_jobs",
        "parameters": [
          {
            "description": "true: only jobs an applicant in T\u00fcrkiye can apply to; false: only those they can't.",
            "in": "query",
            "name": "open_to_you",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "true: only jobs an applicant in T\u00fcrkiye can apply to; false: only those they can't.",
              "title": "Open To You"
            }
          },
          {
            "description": "Only jobs with a match_score of at least this (0-100); unscored jobs drop out.",
            "in": "query",
            "name": "min_score",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100,
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only jobs with a match_score of at least this (0-100); unscored jobs drop out.",
              "title": "Min Score"
            }
          },
          {
            "description": "Only jobs that ask for this skill, e.g. Python (case-insensitive).",
            "in": "query",
            "name": "skill",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 60,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only jobs that ask for this skill, e.g. Python (case-insensitive).",
              "title": "Skill"
            }
          },
          {
            "description": "Only jobs with this work type.",
            "in": "query",
            "name": "work_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/WorkType"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only jobs with this work type.",
              "title": "Work Type"
            }
          },
          {
            "description": "Only jobs from this source.",
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Source"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only jobs from this source.",
              "title": "Source"
            }
          },
          {
            "description": "Only jobs first seen on or after this date (YYYY-MM-DD).",
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only jobs first seen on or after this date (YYYY-MM-DD).",
              "title": "Since"
            }
          },
          {
            "description": "Text to find in the title or company (case-insensitive).",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Text to find in the title or company (case-insensitive).",
              "title": "Q"
            }
          },
          {
            "description": "Page size (1-100).",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Page size (1-100).",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "How many jobs to skip, for paging (0-10000).",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "How many jobs to skip, for paging (0-10000).",
              "maximum": 10000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List jobs, best CV match first, with optional filters"
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "description": "Returns every stored field of a job, including the full description, restrictions, red flags and each skill with whether the owner already has it (on_cv).",
        "operationId": "get_job",
        "parameters": [
          {
            "description": "Job id from list_jobs.",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "description": "Job id from list_jobs.",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Job Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "No job with this id"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get one job with its full description and skills"
      }
    },
    "/runs": {
      "get": {
        "description": "One entry per load into the database. Use it to check whether today's daily load landed: the newest daily entry should have today's run_date.",
        "operationId": "recent_runs",
        "parameters": [
          {
            "description": "How many loads to return (1-50).",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "description": "How many loads to return (1-50).",
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Latest database loads, newest first"
      }
    },
...
  }
}